requirements.txt
aiohttp==3.9.5
bcrypt==4.1.3
fastapi==0.110.2
mysql-connector-python==8.4.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
redis==5.0.7
shortuuid==1.0.13
uvicorn==0.29.0

